+++ test/python/simulators/test_extensions_simulator.py
@@ -13,7 +13,7 @@ import unittest
 import numpy as np
 import qiskit
 import qiskit.extensions.simulator
-from qiskit import BasicAer, LegacySimulators
+from qiskit import BasicAer, Aer
 from qiskit.quantum_info import state_fidelity
 from qiskit.result.postprocess import format_statevector
 from qiskit import execute
@@ -39,7 +39,7 @@ class TestExtensionsSimulator(QiskitTestCase):
         circuit.h(qr[1])
         circuit.load(1)
 
-        sim = LegacySimulators.get_backend('statevector_simulator')
+        sim = Aer.get_backend('statevector_simulator')
         result = execute(circuit, sim).result()
         statevector = result.get_statevector()
         target = [0.70710678 + 0.j, 0.70710678 + 0.j, 0. + 0.j, 0. + 0.j]
@@ -59,7 +59,7 @@ class TestExtensionsSimulator(QiskitTestCase):
         circuit.cx(qr[0], qr[1])
         circuit.h(qr[1])
 
-        sim = LegacySimulators.get_backend('statevector_simulator')
+        sim = Aer.get_backend('statevector_simulator')
         result = execute(circuit, sim).result()
         # TODO: rely on Result.get_statevector() postprocessing rather than manual
         snapshots = result.data(0)['snapshots']['statevector']['3']
@@ -87,7 +87,7 @@ class TestExtensionsSimulator(QiskitTestCase):
                 'id': {'p_pauli': [1.0, 0.0, 0.0]}
             }
         }
-        sim = LegacySimulators.get_backend('qasm_simulator')
+        sim = Aer.get_backend('qasm_simulator')
         shots = 1000
         result = execute(circuit, sim, config=config, shots=shots).result()
         counts = result.get_counts()